pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "tango-da"
version = "2.0"
authors = [
    {name = "Rafael Gort", email = "rafael.gort@inficon.com"},
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest",
    "pytango",
    "h5py",
    "numpy",
]
classifiers=[
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Unix",
]
keywords= ["lib", "data-acquisition", "data-aggregator"]